Equivalence test fixes

+++ ApogeeVGC/Sim/Actions/MoveAction.cs
@@ -44,9 +44,9 @@ public record MoveAction : IAction
             }
         }
     }
-    public int Priority { get; set; }
-    public int FractionalPriority { get; set; }
-    public int Speed { get; set; }
+    public double Priority { get; set; }
+    public double FractionalPriority { get; set; }
+    public double Speed { get; set; }
     public required Pokemon Pokemon { get; set; }
     public int TargetLoc { get; set; }
     public Pokemon? OriginalTarget { get; set; }
