+++ ApogeeVGC/Sim/Actions/SwitchAction.cs
@@ -41,8 +41,8 @@ public record SwitchAction : IAction
             }
         }
     }
-    public int Priority { get; init; }
-    public int Speed { get; set; }
+    public double Priority { get; init; }
+    public double Speed { get; set; }
     public required Pokemon Pokemon { get; init; }
     public required Pokemon Target { get; init; }
     public IEffect? SourceEffect { get; init; }
