+++ ApogeeVGC/Sim/Actions/TeamAction.cs
@@ -21,8 +21,8 @@ public record TeamAction : IAction
         }
     }
 
-    public int Priority { get; init; }
-    public int Speed => 1;
+    public double Priority { get; init; }
+    public double Speed => 1;
     public required Pokemon Pokemon { get; init; }
     public int Index { get; init; }
