+++ ApogeeVGC/Sim/BattleClasses/IPriorityComparison.cs
@@ -5,8 +5,8 @@ namespace ApogeeVGC.Sim.BattleClasses;
 public interface IPriorityComparison
 {
     IntFalseUnion Order { get; }
-    int Priority { get; }
-    int Speed { get; }
+    double Priority { get; }
+    double Speed { get; }
     int SubOrder { get; }
     int EffectOrder { get; }
     //EffectHolder? EffectHolder { get; }
