+++ ApogeeVGC/Sim/Events/EventListener.cs
@@ -6,8 +6,8 @@ namespace ApogeeVGC.Sim.Events;
 public class EventListener : EventListenerWithoutPriority, IPriorityComparison
 {
     public IntFalseUnion Order { get; set; } = int.MaxValue;
-    public int Priority { get; set; }
-    public int Speed { get; set; }
+    public double Priority { get; set; }
+    public double Speed { get; set; }
     public int SubOrder { get; set; }
     public int EffectOrder { get; set; }
 }
